+++ CHANGELOG.md
@@ -31,3 +31,9 @@ Aditionally:
 - Increases the chance of finding the application icon
 - Adds bindable mouse controls
 - Improves performance (refresh time down from 5-9ms to 1ms or less)
+
+# Version 14
+- Action 'open app' will focus to the last window/workspace if the focused
+  window is the player window
+- Action 'play/pause' will send a 'Stop' signal to the player if the player is
+  active and 'Pause' isn't available